+++ src/NexJob/Internal/JobExecutor.cs
@@ -1,8 +1,5 @@
-using System.Collections.Concurrent;
 using System.Diagnostics;
-using System.Linq.Expressions;
 using System.Reflection;
-using System.Text.Json;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NexJob.Storage;
@@ -16,11 +13,8 @@ namespace NexJob.Internal;
 /// </summary>
 internal sealed class JobExecutor
 {
-    // Compiled invoker cache: avoids repeated reflection on hot paths
-    private static readonly ConcurrentDictionary<(Type Job, Type Input), Func<object, object, CancellationToken, Task>>
-        InvokerCache = new();
-
     private readonly IJobStorage _storage;
+    private readonly IJobInvokerFactory _invokerFactory;
     private readonly IServiceScopeFactory _scopeFactory;
     private readonly ThrottleRegistry _throttleRegistry;
     private readonly NexJobOptions _options;
@@ -31,6 +25,7 @@ internal sealed class JobExecutor
     /// Initializes a new instance of the <see cref="JobExecutor"/> class.
     /// </summary>
     /// <param name="storage">The job storage.</param>
+    /// <param name="invokerFactory">The job invoker factory.</param>
     /// <param name="scopeFactory">The scope factory.</param>
     /// <param name="throttleRegistry">The throttle registry.</param>
     /// <param name="options">The nex job options.</param>
@@ -38,6 +33,7 @@ internal sealed class JobExecutor
     /// <param name="logger">The logger.</param>
     public JobExecutor(
         IJobStorage storage,
+        IJobInvokerFactory invokerFactory,
         IServiceScopeFactory scopeFactory,
         ThrottleRegistry throttleRegistry,
         NexJobOptions options,
@@ -45,6 +41,7 @@ internal sealed class JobExecutor
         ILogger<JobExecutor> logger)
     {
         _storage = storage;
+        _invokerFactory = invokerFactory;
         _scopeFactory = scopeFactory;
         _throttleRegistry = throttleRegistry;
         _options = options;
@@ -78,7 +75,7 @@ internal sealed class JobExecutor
 
         try
         {
-            using var context = await PrepareInvocationAsync(job).ConfigureAwait(false);
+            using var context = await _invokerFactory.PrepareAsync(job, cts.Token).ConfigureAwait(false);
             await ExecuteWithThrottlingAndFiltersAsync(context, job, cts.Token).ConfigureAwait(false);
 
             sw.Stop();
@@ -128,50 +125,6 @@ internal sealed class JobExecutor
             new TagList { { "nexjob.job_type", jobType } });
     }
 
-    private static Func<object, object, CancellationToken, Task> GetOrBuildInvoker(
-        Type jobType, Type inputType)
-    {
-        return InvokerCache.GetOrAdd((jobType, inputType), static key =>
-        {
-            var (jt, it) = key;
-
-            if (it == typeof(NoInput))
-            {
-                var noInputMethod = jt.GetMethod(
-                    nameof(IJob.ExecuteAsync),
-                    [typeof(CancellationToken)])!;
-
-                var jobParam = Expression.Parameter(typeof(object), "job");
-                var inputParam = Expression.Parameter(typeof(object), "input"); // ignored
-                var ctParam = Expression.Parameter(typeof(CancellationToken), "ct");
-
-                var call = Expression.Call(
-                    Expression.Convert(jobParam, jt),
-                    noInputMethod,
-                    ctParam);
-
-                return Expression.Lambda<Func<object, object, CancellationToken, Task>>(
-                    call, jobParam, inputParam, ctParam).Compile();
-            }
-
-            var method = jt.GetMethod(nameof(IJob<object>.ExecuteAsync),
-                [it, typeof(CancellationToken)])!;
-
-            var jp = Expression.Parameter(typeof(object), "job");
-            var ip = Expression.Parameter(typeof(object), "input");
-            var ct = Expression.Parameter(typeof(CancellationToken), "ct");
-
-            var callTyped = Expression.Call(
-                Expression.Convert(jp, jt),
-                method,
-                Expression.Convert(ip, it),
-                ct);
-
-            return Expression.Lambda<Func<object, object, CancellationToken, Task>>(
-                callTyped, jp, ip, ct).Compile();
-        });
-    }
-
     private async Task<bool> TryHandleExpirationAsync(JobRecord job)
     {
         if (!job.ExpiresAt.HasValue || DateTimeOffset.UtcNow <= job.ExpiresAt.Value)
@@ -188,33 +141,6 @@ internal sealed class JobExecutor
         NexJobMetrics.JobsExpired.Add(1, new TagList { { "nexjob.job_type", job.JobType } });
 
         return true;
-    }
-
-    private async Task<JobInvocationContext> PrepareInvocationAsync(JobRecord job)
-    {
-        var scope = _scopeFactory.CreateScope();
-
-        scope.ServiceProvider.GetRequiredService<IJobContextAccessor>().Context =
-            new JobContext(job, _storage);
-
-        var jobType = JobTypeResolver.ResolveJobType(job.JobType)
-                      ?? throw new InvalidOperationException($"Cannot load job type: {job.JobType}");
-        var inputType = JobTypeResolver.ResolveInputType(job.InputType)
-                       ?? throw new InvalidOperationException($"Cannot load input type: {job.InputType}");
-
-        var currentVersion = jobType.GetCustomAttribute<SchemaVersionAttribute>()?.Version ?? 1;
-        var migratedJson = scope.ServiceProvider
-            .GetRequiredService<IMigrationPipeline>()
-            .Migrate(job.InputJson, job.SchemaVersion, currentVersion, inputType);
-
-        var input = JsonSerializer.Deserialize(migratedJson, inputType)
-                    ?? throw new InvalidOperationException($"Deserialized null input for job {job.Id}.");
-
-        var jobInstance = scope.ServiceProvider.GetRequiredService(jobType);
-        var invoker = GetOrBuildInvoker(jobType, inputType);
-        var throttleAttrs = jobType.GetCustomAttributes<ThrottleAttribute>(inherit: true);
-
-        return await Task.FromResult(new JobInvocationContext(scope, jobInstance, input, invoker, throttleAttrs)).ConfigureAwait(false);
     }
 
     private async Task ExecuteWithThrottlingAndFiltersAsync(
@@ -381,15 +307,5 @@ internal sealed class JobExecutor
         {
             // Expected on job completion
         }
-    }
-
-    private sealed record JobInvocationContext(
-        IServiceScope Scope,
-        object JobInstance,
-        object Input,
-        Func<object, object, CancellationToken, Task> Invoker,
-        IEnumerable<ThrottleAttribute> ThrottleAttributes) : IDisposable
-    {
-        public void Dispose() => Scope.Dispose();
     }
 }
