+++ src/NexJob/NexJobServiceCollectionExtensions.cs
@@ -173,6 +173,7 @@ public static class NexJobServiceCollectionExtensions
         services.AddSingleton<JobWakeUpChannel>();
         services.AddSingleton<IScheduler, DefaultScheduler>();
         services.AddSingleton(sp => new ThrottleRegistry(sp.GetService<IDistributedThrottleStore>()));
+        services.AddSingleton<IJobInvokerFactory, DefaultJobInvokerFactory>();
         services.AddSingleton<JobExecutor>();
         services.AddSingleton<JobCaptureLoggerProvider>();
         services.AddSingleton<ILoggerProvider>(sp => sp.GetRequiredService<JobCaptureLoggerProvider>());
